# docs/swagger.yaml
definitions:
...
  request.GetName:
    properties:
      forced:
        type: boolean
      username:
        type: string
    type: object
  request.Login:
    properties:
      captchaId:
        type: string
      captchaValue:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  request.NewPassword:
    properties:
      newPassword:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  request.PodRequest:
    properties:
      containerName:
        type: string
      containerPort:
        type: integer
      image:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      namespace:
        type: string
      podName:
        type: string
    type: object
  request.PodUpdateMessage:
    properties:
      name:
        type: string
      namespace:
        type: string
      newImage:
        type: string
    type: object
  response.CommonResponse:
    properties:
      data: {}
      msg:
        type: string
    type: object
  system.User:
    properties:
      createdAt:
        description: 创建时间
        type: string
      email:
        type: string
      id:
        description: 主键ID
        type: integer
      nickName:
        type: string
      password:
        type: string
      phone:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is a k8s cluster management service
  title: K8S Client API
  version: 0.0.1
paths:
...
  /base/login:
    post:
      parameters:
      - description: 用户名，密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: 用户登录
      tags:
      - Base
...
  /pod/create:
    post:
      parameters:
      - description: Pod simple configuration
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建 Pod
      tags:
      - Pod
  /pod/delete:
    delete:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      - description: pod名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除单个 Pod
      tags:
      - Pod
  /pod/get:
    get:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      - description: pod名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取单个 Pod 信息
      tags:
      - Pod
  /pod/list:
    get:
      parameters:
      - default: default
        description: 命名空间
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取命名空间中所有的  Pod 信息
      tags:
      - Pod
  /pod/update:
    put:
      parameters:
      - description: Pod configuration information that needs to be changed
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PodUpdateMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新 Pod
      tags:
      - Pod
  /user/create:
    post:
      parameters:
      - description: 用户名，密码，昵称，手机和邮箱
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 注册用户
      tags:
      - User
  /user/delete:
    delete:
      parameters:
      - description: 用户名，密码，昵称，手机和邮箱
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - User
  /user/get:
    get:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - User
  /user/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 列出所有用户
      tags:
      - User
  /user/update:
    put:
      parameters:
      - description: 用户名，原密码，新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.NewPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
      security:
      - ApiKeyAuth: []
      summary: 用户更新密码
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"